chore(compass-components): Misc element editor fixes
- Use css to toggle element actions (hooks are too slow for this use-case)
- Limit min height of the text area by max height
- Add spacing for type selector
- Move Element to its own file
