+++ packages/compass-components/src/components/document-list/auto-focus-context.tsx
@@ -1,0 +1,12 @@
+import { createContext, useContext } from 'react';
+
+type AutoFocusInfo = {
+  id: string;
+  type: 'key' | 'value';
+} | null;
+
+export const AutoFocusContext = createContext<AutoFocusInfo>(null);
+
+export function useAutoFocusContext(): AutoFocusInfo {
+  return useContext(AutoFocusContext);
+}
