+++ packages/compass-components/src/components/document-list/use-force-update.tsx
@@ -1,0 +1,9 @@
+import { useCallback, useState } from 'react';
+
+export function useForceUpdate(): () => void {
+  const [, setState] = useState({});
+  const forceUpdate = useCallback(() => {
+    setState({});
+  }, []);
+  return forceUpdate;
+}
